learn/requirements.txt
aiosqlite==0.19.0
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
apprise==1.6.0
asgi-lifespan==2.1.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coolname==2.2.0
croniter==2.0.1
cryptography==41.0.7
dateparser==1.2.0
dnspython==2.4.2
docker==6.1.3
email-validator==2.1.0.post1
fsspec==2023.12.2
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-spreadsheet==0.0.6
googleapis-common-protos==1.62.0
graphviz==0.20.1
greenlet==3.0.2
griffe==0.38.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kubernetes==28.1.0
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.2
oauth2client==4.1.3
oauthlib==3.2.2
orjson==3.9.10
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
pendulum==2.1.2
prefect==2.14.11
prefect-docker==0.4.1
prefect-kubernetes==0.3.2
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
readchar==4.0.5
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.15.2
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.32.0.post1
text-unidecode==1.3
toml==0.10.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ujson==5.9.0
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.24.0.post1
websocket-client==1.7.0
websockets==12.0
